Fix #2 by trimming the stdlib to between both { }, not just {.
Brython's stdlib changed format since this was written and now
contains extra data at the end of the JSON data.
This means that trimming the contents of this file to just the
leading { is not enough, we must also find the closing } and
trim to that.
Updated the dummy stdlib file in tests.
With the fix in place, tests once again pass with Brython 3.9.5.